Submission
I actually feel like this solution is too hard coded, but it works.

I've been messin around with other solutions for about 14 hours. They become very messy and didn't get to an end. But a I learned sooo much just playing around, finding solutions for problems that seemed to not make any sense at first and so on. I really got into the flow of programming, I guess, forgetting everything around me. Was a very nice experience :)

I think the expectation here was not to create a fully modular program. So I got back to the easier solution. Like I said, i think it works and that's probably what counts here.

Thanks for the amazing course by the way!  :)